Backend/requirement.txt
fastapi==0.111.1
uvicorn==0.35.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
pydantic==2.6.0